+++ ext/raptor_native/extconf.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require "mkmf"
+
+append_cflags("-fvisibility=hidden")
+
+create_makefile("raptor/raptor_native")